requirements.txt
# Django
Django==2.0.3
django-braces==1.12.0
django-simple-menu==1.2.1
django-rosetta==0.8.1

# django-modeladmin-reorder==0.2
# Django 2 support is only available in master
-e git+https://github.com/mishbahr/django-modeladmin-reorder.git@fb40efa6d3d0668f3323aaca8e5c4486c0ffdd7b#egg=django-modeladmin-reorder


python-dateutil==2.6.1
mysqlclient==1.3.12
raven==6.6.0
InstagramAPI==1.0.2
requests==2.18.4
